amer_Accountant

+++ database/migrations/2023_05_06_214241_create_fees_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('fees', function (Blueprint $table) {
             $table->id();
             $table->string('title', 100);
-            $table->decimal('amount',8,2);
+            $table->bigInteger('amount');
             $table->foreignId('Grade_id')->references('id')->on('Grades')->onDelete('cascade');
             $table->foreignId('Classroom_id')->references('id')->on('Classrooms')->onDelete('cascade');
             $table->string('description')->nullable();
