+++ database/migrations/2023_05_22_183135_create_fund_accounts_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->date('date');
             $table->foreignId('receipt_id')->nullable()->references('id')->on('receipt_students')->onDelete('cascade');
             $table->foreignId('payment_id')->nullable()->references('id')->on('payment_students')->onDelete('cascade');
-            $table->decimal('Debit',8,2)->nullable();
-            $table->decimal('credit',8,2)->nullable();
+            $table->bigInteger('Debit')->nullable();
+            $table->bigInteger('credit')->nullable();
             $table->string('description');
             $table->timestamps();
         });
